// WebApplicationPlateforme/ClientApp/src/app/shared/Models/Taches/tache.model.ts
export class Tache {
  id?: number;
  date: string;
  priorite: string;
  tacheNom: string;
  description: string;
  delai: string;
  atous: string;
  createur: string;
  etat: string;
  type: string;
  sousProjet: string;
  Attribut1: string;
  Attribut2: string;
  Attribut3: string;
  attribut4 : number;
  creatorName: string;
  idUserCreator: string;
  affectedName: string;
}